+++ CHANGELOG.md
@@ -91,6 +91,7 @@ count (
 
 ### Bug fixes and improvements
 
+- server: validate saved REST endpoint queries wrt schema
 - server: improved error reporting for env vars in `test_webhook_transform` metadata API endpoint
 - server: add a placeholder field to the schema when the `query_root` would be empty
 - server: fix invalid GraphQL name in the schema arising from a remote relationship from a table in a custom schema